reorganize test directories
