+++ client/__tests__/util/typedCrossfilter/bitArray.test.js
@@ -1,6 +1,6 @@
 // jshint esversion: 6
 
-import BitArray from "../../src/util/typedCrossfilter/bitArray";
+import BitArray from "../../../src/util/typedCrossfilter/bitArray";
 
 const defaultTestLength = 8;
