+++ client/__tests__/util/typedCrossfilter/positiveInterval.test.js
@@ -1,7 +1,7 @@
 // jshint esversion: 6
 
 // const PositiveIntervals = require("../../src/util/typedCrossfilter/positiveIntervals");
-import PositiveIntervals from "../../src/util/typedCrossfilter/positiveIntervals";
+import PositiveIntervals from "../../../src/util/typedCrossfilter/positiveIntervals";
 
 describe("canonicalize", () => {
   test("empty", () => {
